Started todays session by finishing the stratified sampling implementation of the books example. Afterwards i tested how it would affect the sampling bias on the dataset's representation. Stratified sampling rendered better results in both bias error and in data visualization than te previously imlpemented random sampling.
Some details of the stratified sampling technique remain unclear to me.
I need to reread stratified sampling as well as take some time
understanding the current implementation. So far i understand that the
strata is defined by the income parameter which is a column derived from
the dataset. On line 119 i used a native method from sklearn which is
called StratifiedShuffleSplit which does what the names says in order to
generate the indexing splitting criteria. Then i used a for loop which
takes said splitter into account for actually splitting the current
dataset into the desired test and train sets.

Finally i continued by representing the data into a colormap which for
some reason i still do not comprehend it cant be defined as a jet
colormap but if i leave it as it is it just works but with a different
color scheme.